take ScaleEdges and ScaleAreas out of WarpX class
