+++ Source/Initialization/WarpXInitData.cpp
@@ -17,6 +17,9 @@
 #include "Diagnostics/MultiDiagnostics.H"
 #include "Diagnostics/ReducedDiags/MultiReducedDiags.H"
 #include "EmbeddedBoundary/Enabled.H"
+#ifdef AMREX_USE_EB
+#   include "EmbeddedBoundary/EmbeddedBoundary.H"
+#endif
 #include "Fields.H"
 #include "FieldSolver/ElectrostaticSolvers/ElectrostaticSolver.H"
 #include "FieldSolver/FiniteDifferenceSolver/MacroscopicProperties/MacroscopicProperties.H"
@@ -1237,11 +1240,11 @@ void WarpX::InitializeEBGridData (int lev)
 
                 auto edge_lengths_lev = m_fields.get_alldirs(FieldType::edge_lengths, lev);
                 ComputeEdgeLengths(edge_lengths_lev, eb_fact);
-                ScaleEdges(edge_lengths_lev, CellSize(lev));
+                warpx::embedded_boundary::ScaleEdges(edge_lengths_lev, CellSize(lev));
 
                 auto face_areas_lev = m_fields.get_alldirs(FieldType::face_areas, lev);
                 ComputeFaceAreas(face_areas_lev, eb_fact);
-                ScaleAreas(face_areas_lev, CellSize(lev));
+                warpx::embedded_boundary::ScaleAreas(face_areas_lev, CellSize(lev));
 
                 // Compute additional quantities required for the ECT solver
                 MarkExtensionCells();
